feat: refactor logic by platform
