+++ gsd-lite/ARCHITECTURE.md
@@ -4,53 +4,56 @@
 
 ## Project Structure Overview
 
-| File | Purpose |
-|------|---------|
-| `index.html` | Main application entry point; defines UI and loads scripts |
-| `clipboard2markdown.js` | **Controller**: Handles user input, clipboard events, and orchestrates conversion |
-| `to-markdown.js` | **Engine**: The core library that parses HTML and converts it to Markdown |
-| `bootstrap.css` | Styling framework for the UI |
-| `background.svg` | Background asset |
+| File / Directory | Purpose |
+|---|---|
+| `index.html` | Main application entry point. Loads the main script as an ES Module. |
+| `clipboard2markdown.js` | **UI Controller**: Handles DOM events (paste, keydown) and orchestrates the conversion process. Imports logic from `src/platforms`. |
+| `vite.config.js` | Vite build configuration. Defines the output directory (`dist/`) and base path for deployment. |
+| `package.json` | Project metadata and dependencies (`vite`, `turndown`). Defines `dev`, `build`, and `test` scripts. |
+| `src/platforms/` | **Core Logic**: Contains all platform-specific conversion logic. |
+| `src/platforms/jira.js` | Exports Jira-specific Turndown `rules` and a `sanitizer` function for pre-processing HTML. |
+| `src/platforms/common.js` | Exports rules and sanitizers that apply to all sources. |
+| `src/platforms/index.js` | **Aggregator**: Imports from all other platform files and exports `addAllRules` and `sanitize` helpers. |
+| `lib/` | Contains the core `turndown` and `turndown-plugin-gfm` libraries. |
+| `dist/` | **Build Output**: Contains the minified and optimized assets for production deployment. This directory is generated by `npm run build`. |
+| `tests/` | Contains test fixtures and the automated test runner script. |
 
 ## Tech Stack
 
+- **Build Tool:** Vite (provides a dev server with HMR and builds optimized assets for production).
+- **Package Manager:** npm
 - **Runtime:** Browser (Client-side JavaScript)
-- **Language:** HTML5, CSS3, JavaScript (ES5+)
+- **Language:** HTML5, CSS3, JavaScript (ES Modules)
 - **Key Dependencies:**
-  - `to-markdown.js` - The core conversion engine (fork/embedded version of domchristie/to-markdown)
-  - `clipboard2markdown.js` - Application logic and custom converter rules (Pandoc-style)
+  - `turndown`: The core HTML-to-Markdown conversion engine.
+  - `turndown-plugin-gfm`: Provides GFM extensions (tables, strikethrough).
+  - `vitest` / `jsdom` (for testing): Enables automated testing in a Node.js environment.
 
-## End-to-End Flow
-
-### Interaction Diagram
+## End-to-End Flow (Conversion Logic)
 
 ```mermaid
 sequenceDiagram
     participant User
     participant Browser
-    participant Controller as clipboard2markdown.js
-    participant DOM as Hidden DOM (#pastebin)
-    participant Engine as to-markdown.js
-    participant UI as Output (#output)
+    participant UIController as clipboard2markdown.js
+    participant Aggregator as src/platforms/index.js
+    participant Sanitizers as Jira, Common Sanitizers
+    participant Engine as Turndown.js
+    participant Rules as Jira, Common Rules
 
-    User->>Browser: Press Ctrl+V / Cmd+V
-    Browser->>Controller: Keydown Event (Global)
-    Controller->>DOM: Focus & Clear #pastebin
-    User->>DOM: Paste Content (Native Browser Action)
-    DOM-->>Controller: Paste Event Triggered
-    Controller->>Controller: Wait 200ms (setTimeout) to allow DOM update
-    Controller->>DOM: Read innerHTML
-    DOM-->>Controller: HTML String
-    Controller->>Engine: convert(html)
-    loop Engine Processing
-        Engine->>Engine: Parse HTML to DOM Tree
-        Engine->>Engine: Flatten to Array (BFS)
-        Engine->>Engine: Iterate Reverse (Deepest First)
-        Engine->>Engine: Apply Converters
+    User->>Browser: Pastes HTML content
+    Browser->>UIController: 'paste' event triggered
+    UIController->>UIController: Reads pasted innerHTML
+    UIController->>Aggregator: Calls master sanitize(html)
+    Aggregator->>Sanitizers: Runs each platform's sanitizer function on the DOM
+    Sanitizers-->>Aggregator: Mutated DOM is ready
+    Aggregator-->>UIController: Returns sanitized HTML string
+    UIController->>Engine: turndown(sanitizedHtml)
+    loop Turndown Processing
+        Engine->>Rules: Matches nodes against all registered platform rules
     end
-    Engine-->>Controller: Markdown String
-    Controller->>UI: Insert Markdown & Select Text
-    UI-->>User: Display Result
+    Engine-->>UIController: Final Markdown String
+    UIController->>User: Displays Markdown in output
 ```
 
 ## Logic Outline
@@ -96,8 +99,7 @@ This library handles the heavy lifting of parsing and traversing the HTML.
 
 ## Entry Points
 
-Start reading here:
-
--   `clipboard2markdown.js` - **Read `convert()` first**: This is the high-level function that ties everything together. Then look at the `pandoc` array to see the specific transformation rules.
--   `index.html` - **View Source**: See the `div#pastebin` and how scripts are loaded.
--   `to-markdown.js` - **Read `toMarkdown` function**: Understand the parsing -> flattening -> reverse-iteration pipeline.
+-   **`package.json`**: Look at the `scripts` to understand how to run the dev server (`dev`), build (`build`), and test (`test`).
+-   **`clipboard2markdown.js`**: Read the `paste` event handler to see the main orchestration logic.
+-   **`src/platforms/index.js`**: Understand how platform-specific modules are aggregated and applied.
+-   **`src/platforms/jira.js`**: See a concrete example of a platform-specific `sanitizer` and `ruleset`.
